MuploadJSON/requirements.txt
mstrio-py
pandas